requirements.txt
anthropic>=0.18.0
pyyaml>=6.0
